requirements.txt
pydraughts==0.5.7
PyYAML>=5.4.1
requests>=2.25.1
backoff==2.2.1
rich==13.6.0
cityhash

# Requirements for tests
pytest==7.4.2
pytest-timeout==2.2.0
